Episode 001 - Image Gallery
This is a simple image gallery made in Python with PyQt5.  It shows the image, can be paused and resumed, can have the time adjusted to switch images, has a next button, and a previous button.  It also has a menu with a donate and about option.  Make it your own and have fun!

Check out more videos at https://www.youtube.com/@yourpststudios